# openapi/altus4.openapi.yaml
openapi: 3.0.3
info:
  title: Altus4 Core API
  version: 0.4.0
  description: |
    Complete API specification for Altus4 - AI-Enhanced MySQL Full-Text Search Engine.

    ## Authentication

    Altus4 uses dual authentication:
    - **JWT Bearer Tokens**: For user management, API key management, database management, and analytics
    - **API Key Bearer Tokens**: For search operations and B2B service integration

    API Key format: `altus4_sk_{env}_{secret}` where env is 'test' or 'live'.
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    apiKeyAuth:
      type: http
      scheme: bearer
      description: API key for search operations (format altus4_sk_{env}_{secret})
  schemas:
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        data:
          description: Response data (varies by endpoint)
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code for programmatic handling
            message:
              type: string
              description: Human-readable error message
            details:
              description: Additional error details
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
            version:
              type: string
            executionTime:
              type: number
              description: Request execution time in milliseconds
            apiKeyTier:
              type: string
              description: Present on API key authenticated responses
...
  responses:
...
    Unauthorized:
      description: Unauthorized - Missing or invalid token
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: INVALID_API_KEY
                      message:
                        type: string
                        example: Missing or invalid API key
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: INSUFFICIENT_PERMISSIONS
                      message:
                        type: string
                        example: API key lacks required permissions
...
paths:
...
  /api/v1/analytics/search-trends:
    get:
      tags: [Analytics]
      summary: Get search trends analytics
      description: Get detailed search trends and analytics data
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date
          description: Start date for the analytics period
        - in: query
          name: endDate
          schema:
            type: string
            format: date
          description: End date for the analytics period
        - in: query
          name: period
          schema:
            type: string
            enum: [day, week, month, 3months, 6months, year]
            default: week
          description: Analytics period grouping
      responses:
        '200':
          description: Search trends analytics retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            period:
                              type: string
                              enum: [day, week, month, 3months, 6months, year]
                            topQueries:
                              type: array
                              items:
                                type: string
                            queryVolume:
                              type: number
                            avgResponseTime:
                              type: number
                            popularCategories:
                              type: array
                              items:
                                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/analytics/performance:
    get:
      tags: [Analytics]
      summary: Get performance metrics
      description: Get detailed performance metrics and analytics
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
        - in: query
          name: period
          schema:
            type: string
            enum: [day, week, month, 3months, 6months, year]
            default: week
      responses:
        '200':
          description: Performance metrics retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          timeSeriesData:
                            type: array
                            items:
                              type: object
                              properties:
                                date:
                                  type: string
                                  format: date
                                avg_response_time:
                                  type: number
                                max_response_time:
                                  type: number
                                min_response_time:
                                  type: number
                                total_queries:
                                  type: integer
                                avg_results:
                                  type: number
                          searchModeDistribution:
                            type: array
                            items:
                              type: object
                              properties:
                                search_mode:
                                  type: string
                                count:
                                  type: integer
                                avg_time:
                                  type: number
                          summary:
                            type: object
                            properties:
                              totalQueries:
                                type: integer
                              averageResponseTime:
                                type: number
                              averageResults:
                                type: number
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/analytics/popular-queries:
    get:
      tags: [Analytics]
      summary: Get popular queries
      description: Get the most popular search queries
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
        - in: query
          name: period
          schema:
            type: string
            enum: [day, week, month, 3months, 6months, year]
            default: week
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Popular queries retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            query_text:
                              type: string
                            frequency:
                              type: integer
                            avg_time:
                              type: number
                            avg_results:
                              type: number
                            last_used:
                              type: string
                              format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/analytics/search-history:
    get:
      tags: [Analytics]
      summary: Get detailed search history
      description: Get detailed search history with analytics data
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Search history retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              type: object
                              properties:
                                id:
                                  type: string
                                query:
                                  type: string
                                database:
                                  type: string
                                resultCount:
                                  type: integer
                                executionTime:
                                  type: number
                                timestamp:
                                  type: string
                                  format: date-time
                          total:
                            type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/analytics/insights:
    get:
      tags: [Analytics]
      summary: Get AI-generated insights
      description: Get AI-generated insights about search patterns and performance
      security:
        - bearerAuth: []
      responses:
        '200':
          description: AI insights retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            type:
                              type: string
                              enum: [query_optimization, trend_analysis, semantic_enhancement]
                            confidence:
                              type: number
                            description:
                              type: string
                            actionable:
                              type: boolean
                            data:
                              type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/analytics/dashboard:
    get:
      tags: [Analytics]
      summary: Get dashboard data
      description: Get comprehensive dashboard data for analytics overview
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard data retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          trends:
                            type: object
                          performance:
                            type: object
                          popularQueries:
                            type: array
                            items:
                              type: object
                          insights:
                            type: array
                            items:
                              type: object
                          summary:
                            type: object
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Admin Analytics Endpoints
  /api/v1/analytics/admin/system-overview:
    get:
      tags: [Analytics]
      summary: Get system overview (Admin)
      description: Get system-wide overview data (admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: System overview retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          totalUsers:
                            type: integer
                          totalSearches:
                            type: integer
                          totalDatabases:
                            type: integer
                          systemHealth:
                            type: object
                            properties:
                              status:
                                type: string
                              uptime:
                                type: number
                              memoryUsage:
                                type: number
                              cpuUsage:
                                type: number
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/analytics/admin/user-activity:
    get:
      tags: [Analytics]
      summary: Get user activity (Admin)
      description: Get user activity data across the system (admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User activity data retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          activeUsers:
                            type: integer
                          newUsers:
                            type: integer
                          topUsers:
                            type: array
                            items:
                              type: object
                              properties:
                                userId:
                                  type: string
                                email:
                                  type: string
                                searchCount:
                                  type: integer
                                lastActive:
                                  type: string
                                  format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...